API Testing
commit 83654687dfd2d4d45358e6070990ba19cde16a58
Author: al masud uddin <monicse09ku@yahoo.com>
Date:   Sun Dec 7 16:14:08 2025 +0600

    remove codes

commit e9f149a9dec90af54b338b12727d57e2d86a45b3
Author: al masud uddin <monicse09ku@yahoo.com>
Date:   Sun Dec 7 15:46:47 2025 +0600

    Tasts Testing Addition

commit 9b8fcaf9db4d85acd8bb33f9223aad553743603e
Author: al masud uddin <monicse09ku@yahoo.com>
Date:   Sun Dec 7 15:35:29 2025 +0600

    Projects Testing Addition

commit 9a0c60418f678c95fdcbb949ef92bca5e157961a
Author: al masud uddin <monicse09ku@yahoo.com>
Date:   Sun Dec 7 15:25:45 2025 +0600

    Clients Testing Addition

+++ app/Http/Controllers/Api/ProjectController.php
@@ -53,8 +53,14 @@ class ProjectController extends Controller
     /**
      * Update a project - Only admins can update
      */
-    public function update(ProjectRequest $request, Project $project)
+    public function update(ProjectRequest $request, $id)
     {
+        $project = Project::find($id);
+        
+        if (!$project) {
+            return $this->error('Project not found', null, 404);
+        }
+        
         $this->authorize('update', $project);
 
         $validated = $request->validated();
@@ -68,8 +74,14 @@ class ProjectController extends Controller
     /**
      * Delete a project - Only admins can delete
      */
-    public function destroy(Project $project)
+    public function destroy($id)
     {
+        $project = Project::find($id);
+        
+        if (!$project) {
+            return $this->error('Project not found', null, 404);
+        }
+        
         $this->authorize('delete', $project);
 
         $project->delete();
